--- internal/slack/go.mod
@@ -1,5 +1,0 @@
-module slack
-
-go 1.22.3
-
-require github.com/joho/godotenv v1.5.1 // indirect
