+++ internal/slack/slack.go
@@ -1,4 +1,4 @@
-package main
+package slack
 
 import (
 	"bytes"
